pyproject.toml
[project]
name = "starjaxrl"
version = "0.1.0"
description = "Starship belly-flop landing via JAX-based reinforcement learning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flax>=0.12.5",
    "gymnasium>=1.2.3",
    "hydra-core>=1.3.2",
    "jax>=0.9.1",
    "jaxlib>=0.9.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "optax>=0.2.6",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "wandb>=0.25.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
